Refactor

+++ src/resolvers/author.resolver.ts
@@ -84,11 +84,16 @@ export class AuthorResolver {
    }
 
    @Mutation (()=>Boolean)
-  async deleteOneAuthor(
+   async deleteOneAuthor(
     @Arg ("input",() => AuthorIdInput)input:AuthorIdInput
   ):Promise<Boolean>{
+      try{
+        const author = await this.authorRepository.findOne(input.id);
+        if (!author) throw new Error ('Author does not exist');
         await this.authorRepository.delete(input.id);
         return true;
+    } catch (e) {
+        throw new Error(e.message)
     }
-       
+   }
   }
